# PERF: df.groupby(categorical, sort=False)