- [x] [Tests added and passed](https://pandas.pydata.org/pandas-docs/dev/development/contributing_codebase.html#writing-tests) if fixing a bug or adding a new feature
- [x] All [code checks passed](https://pandas.pydata.org/pandas-docs/dev/development/contributing_codebase.html#pre-commit).
- [x] Added an entry in the latest `doc/source/whatsnew/vX.X.X.rst` file if fixing a bug or adding a new feature.

xref: #48749

```
In [1]: df = pd.DataFrame({"key": pd.Categorical(range(1_000_000 - 1, -1, -1), range(1_000_000))})

In [2]: %timeit df.groupby("key", sort=False)  # PR
66.7 ms ± 1.53 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)

In [9]: %timeit df.groupby("key", sort=False)  # main
586 ms ± 4.15 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
```